# New way of adding histograms together with a merge iterator over buckets