**Note: This is pointing to sparsehistogram branch. Only interesting to @beorn7 for now.**

This implements new way of adding histograms together as said in https://github.com/prometheus/prometheus/pull/9926#discussion_r762973794.

And I have also added a benchmark to compare old and new ways. I am happy that I wrote the benchmark because the new proposed way is very slow compared to current way of adding 1 histogram at a time. This PR is hence just a FYI and **not to be merged**.

**Benchmark comparison:**
```
benchmark                                                 old ns/op     new ns/op     delta
BenchmarkHistogramAdd/numHists=10,maxBuckets=10-8         0.00          0.00          +584.27%
BenchmarkHistogramAdd/numHists=10,maxBuckets=30-8         0.00          0.00          +616.57%
BenchmarkHistogramAdd/numHists=10,maxBuckets=50-8         0.00          0.00          +663.60%
BenchmarkHistogramAdd/numHists=10,maxBuckets=100-8        0.00          0.00          +610.48%
BenchmarkHistogramAdd/numHists=10,maxBuckets=150-8        0.00          0.00          +640.42%
BenchmarkHistogramAdd/numHists=100,maxBuckets=10-8        0.00          0.00          +854.62%
BenchmarkHistogramAdd/numHists=100,maxBuckets=30-8        0.00          0.00          +1169.91%
BenchmarkHistogramAdd/numHists=100,maxBuckets=50-8        0.00          0.00          +931.03%
BenchmarkHistogramAdd/numHists=100,maxBuckets=100-8       0.00          0.00          +1009.01%
BenchmarkHistogramAdd/numHists=100,maxBuckets=150-8       0.00          0.01          +909.20%
BenchmarkHistogramAdd/numHists=1000,maxBuckets=10-8       0.00          0.01          +1064.22%
BenchmarkHistogramAdd/numHists=1000,maxBuckets=30-8       0.00          0.02          +1236.43%
BenchmarkHistogramAdd/numHists=1000,maxBuckets=50-8       0.00          0.03          +1134.73%
BenchmarkHistogramAdd/numHists=1000,maxBuckets=100-8      0.00          0.05          +1114.13%
BenchmarkHistogramAdd/numHists=1000,maxBuckets=150-8      0.01          0.06          +1028.35%
BenchmarkHistogramAdd/numHists=10000,maxBuckets=10-8      0.00          0.08          +1658.94%
BenchmarkHistogramAdd/numHists=10000,maxBuckets=30-8      0.01          0.25          +1916.06%
BenchmarkHistogramAdd/numHists=10000,maxBuckets=50-8      0.02          0.38          +1843.54%
BenchmarkHistogramAdd/numHists=10000,maxBuckets=100-8     0.04          0.64          +1687.98%
BenchmarkHistogramAdd/numHists=10000,maxBuckets=150-8     0.05          5769          +11154192.34%
```

**Raw data (since it's showing 0.00 above)**

Old:
```
BenchmarkHistogramAdd/numHists=10,maxBuckets=10-8         	1000000000	         0.0000089 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=30-8         	1000000000	         0.0000175 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=50-8         	1000000000	         0.0000239 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=100-8        	1000000000	         0.0000477 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=150-8        	1000000000	         0.0000621 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=10-8        	1000000000	         0.0000628 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=30-8        	1000000000	         0.0001419 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=50-8        	1000000000	         0.0002610 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=100-8       	1000000000	         0.0004238 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=150-8       	1000000000	         0.0006416 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=10-8       	1000000000	         0.0005584 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=30-8       	1000000000	         0.001400 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=50-8       	1000000000	         0.002309 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=100-8      	1000000000	         0.003778 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=150-8      	1000000000	         0.005485 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=10-8      	1000000000	         0.004725 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=30-8      	1000000000	         0.01245 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=50-8      	1000000000	         0.01966 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=100-8     	1000000000	         0.03594 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=150-8     	1000000000	         0.05172 ns/op	       0 B/op	       0 allocs/op
```

New:
```
BenchmarkHistogramAdd/numHists=10,maxBuckets=10-8         	1000000000	         0.0000609 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=30-8         	1000000000	         0.0001254 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=50-8         	1000000000	         0.0001825 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=100-8        	1000000000	         0.0003389 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10,maxBuckets=150-8        	1000000000	         0.0004598 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=10-8        	1000000000	         0.0005995 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=30-8        	1000000000	         0.001802 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=50-8        	1000000000	         0.002691 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=100-8       	1000000000	         0.004700 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=100,maxBuckets=150-8       	1000000000	         0.006475 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=10-8       	1000000000	         0.006501 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=30-8       	1000000000	         0.01871 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=50-8       	1000000000	         0.02851 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=100-8      	1000000000	         0.04587 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=1000,maxBuckets=150-8      	1000000000	         0.06189 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=10-8      	1000000000	         0.08311 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=30-8      	1000000000	         0.2510 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=50-8      	1000000000	         0.3821 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=100-8     	1000000000	         0.6426 ns/op	       0 B/op	       0 allocs/op
BenchmarkHistogramAdd/numHists=10000,maxBuckets=150-8     	  174598	      5769 ns/op	     652 B/op	       0 allocs/op
```